### 2020-05-22 #2222 @whedon
[PRE REVIEW]: txshift: Efficient estimation of the causal effects of stochastic interventions in R
**Submitting author:** @nhejazi (<a href="http://orcid.org/0000-0002-7127-2789">Nima Hejazi</a>)
**Repository:** <a href="https://github.com/nhejazi/txshift" target ="_blank">https://github.com/nhejazi/txshift</a>
**Version:** 0.3.3
