### 2020-05-27 241e535
update news, bump version
