### 2020-06-05 #2222 @kyleniemeyer
Hi @nhejazi, I will be trying to find an editor to handle your submission. However, one quick comment: for a non-expert in your field, it's very hard for me to understand much about your software—or even figure out the topic area—based on the introductory paragraph. Could you please edit this to make it easier for a non-expert, at least at the beginning?
