### 2020-06-08 #2222 @marcosvital
Hi, @nhejazi, I will be the editor of your submission.

I agree with @kyleniemeyer comment, as [JOSS documentation](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) states that:

> Begin your paper with a summary of the high-level functionality of your software for a non-specialist reader. Avoid jargon in this section.

At the same time you work on that, we can start assigning reviewers for you submission. We usually ask the authors for suggestions, so let me know if you have anyone in mind that could do this - you can paste the github profile of anyone you think that could do it.
You can also [check this list of people](https://bit.ly/joss-reviewers) that have agreed to review for JOSS, and see if any of them may be suitable to review this submission.

Also, feel free to tag the other author here on this issue if you want to.
