### 2020-06-08 #2222 @nhejazi
Thank you for taking a look through the paper, @marcosvital --- I'll begin the process of editing it to include a high-level summary shortly. Also, thanks for the reminder, taggin @benkeser since this was joint work.

In the meantime, from a quick look through the list of potential reviewers, I think @Tgerke, @kellieotto, @mariacuellar would all make excellent reviewers given their experience with R and backgrounds in statistics/biostatistics, epidemiology, causal inference.
