### 2020-06-08 nhejazi/txshift#52 @nhejazi
High-level intro in JOSS paper
Improve JOSS paper by adding a high-level introduction ("Begin your paper with a summary of the high-level functionality of your software for a non-specialist reader. Avoid jargon in this section)", as per https://github.com/openjournals/joss-reviews/issues/2222
