### 2020-06-08 #2222 @mariacuellar
Dear Marcos,

I am not currently available to review this due to having no childcare
during the pandemic. I appreciate the consideration.

All the best,
Maria

On Mon, Jun 8, 2020 at 2:49 PM Marcos Vital (LEQ-UFAL) <
notifications@github.com> wrote:

> Dear @tgerke <https://github.com/tgerke>, @kellieotto
> <https://github.com/kellieotto> and @mariacuellar
> <https://github.com/mariacuellar>
>
> Would you be willing to review this submission for JOSS?
> We carry out our checklist-driven reviews here in GitHub issues and follow
> these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> This is still a pre-review, once there are enough reviewers, we will move
> on to a new issue where the review will take place.
>
> If you are not able to review, please let us know if you have
> recommendations for other reviewers.
>
