### 2020-06-13 49b67d4
update JOSS paper
