### 2020-06-25 4cd911b
add msm_formula and example to docs
