### 2020-06-27 58ec07c
finish spline MSM + add plotting
