+++ R/msm_vim.R
@@ -1,11 +1,13 @@
-#' Marginal structural model for the causal effects of a grid of interventions
+#' Working marginal structural model for causal effects of an intervention grid
 #'
 #' @details Computes estimates of the counterfactual mean over a grid of shift
-#'  stochastic interventions and fits a marginal structural model to summarize
-#'  the trend through the counterfactual means as a function of the specified
-#'  shift intervention. Provides support for two weighting schemes, may be used
-#'  with either the one-step or TML estimator, and allows the construction of
-#'  marginal or simultaneous confidence intervals.
+#'  stochastic interventions and fits a working marginal structural model to
+#'  summarize the trend through the counterfactual means as a function of the
+#'  specified shift intervention. The working marginal structural model may be
+#'  linear in the shift parameter or piecewise linear with a single knot point.
+#'  Provides support for two weighting schemes, may be used with either of the
+#'  one-step or TML estimators, and also allows the construction of marginal or
+#'  simultaneous confidence intervals.
 #'
 #' @param Y A \code{numeric} vector of the observed outcomes.
 #' @param A A \code{numeric} vector corresponding to a treatment variable. The
@@ -23,9 +25,11 @@
 #'  have played a role in defining the sampling/censoring mechanism (C).
 #' @param delta_grid A \code{numeric} vector giving the individual values of
 #'  the shift parameter used in computing each of the estimates.
-#' @param msm_formula The right-hand-side of a formula describing the marginal
-#' structural model. The single covariate in this model is \code{"delta"}. See
-#' examples. 
+#' @param msm_form A \code{list} specifying the type of working MSM to fit to
+#'  summarize the counterfactual means. The \code{list} has two components:
+#'  (1) \code{"type"}, which may be either "linear" or "piecewise", and (2)
+#'  \code{"knot"}, which, if specified, must be a value in \code{delta_grid}.
+#'  See examples for its use.
 #' @param estimator The type of estimator to be fit, either \code{"tmle"} for
 #'  targeted maximum likelihood estimation or \code{"onestep"} for a one-step
 #'  augmented inverse probability weighted (AIPW) estimator.
@@ -44,7 +48,7 @@
 #'
 #' @importFrom assertthat assert_that
 #' @importFrom data.table as.data.table
-#' @importFrom stats cov qnorm pnorm
+#' @importFrom stats cov confint qnorm pnorm as.formula model.matrix
 #' @importFrom mvtnorm qmvnorm
 #'
 #' @return A \code{list} containing estimates of the individual counterfactual
@@ -57,7 +61,7 @@
 #' n_obs <- 100
 #' W <- as.numeric(replicate(1, rbinom(n_obs, 1, 0.5)))
 #' A <- as.numeric(rnorm(n_obs, mean = 2 * W, sd = 1))
-#' Y <- rbinom(n_obs, 1, plogis(A + W + rnorm(n_obs, mean = 0, sd = 1)))
+#' Y <- rbinom(n_obs, 1, plogis(2 * A - W))
 #' msm <- msm_vimshift(
 #'   W = W, A = A, Y = Y, estimator = "tmle",
 #'   g_fit_args = list(
[…]
