+++ man/bound_propensity.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
+% Please edit documentation in R/bound.R
 \name{bound_propensity}
 \alias{bound_propensity}
 \title{Bound Generalized Propensity Score}
