+++ man/confint.txshift.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
+% Please edit documentation in R/confint.R
 \name{confint.txshift}
 \alias{confint.txshift}
 \title{Confidence Intervals for Counterfactual Mean Under Stochastic Intervention}
