+++ man/bound_precision.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
+% Please edit documentation in R/bound.R
 \name{bound_precision}
 \alias{bound_precision}
 \title{Bound Precision}
