+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: txshift
 Title: Efficient Estimation of the Causal Effects of Stochastic Interventions
-Version: 0.3.3
+Version: 0.3.4
 Authors@R: c(
     person("Nima", "Hejazi", email = "nh@nimahejazi.org",
            role = c("aut", "cre", "cph"),
@@ -40,7 +40,11 @@ Imports:
     hal9001 (>= 0.2.5),
     haldensify (>= 0.0.5),
     sl3 (>= 1.3.7),
-    Rdpack
+    lspline,
+    ggplot2,
+    tibble,
+    latex2exp,
+    Rdpack,
 Suggests:
     testthat,
     knitr,
