+++ DESCRIPTION
@@ -37,8 +37,8 @@ Imports:
     data.table,
     assertthat,
     mvtnorm,
-    hal9001 (>= 0.2.5),
-    haldensify (>= 0.0.5),
+    hal9001 (>= 0.2.6),
+    haldensify (>= 0.0.6),
     sl3 (>= 1.3.7),
     lspline,
     ggplot2,
