### 2020-06-27 b6c5523
remove font size spec in plotting
