+++ NAMESPACE
@@ -16,14 +16,12 @@ importFrom(data.table,data.table)
 importFrom(data.table,set)
 importFrom(data.table,setnames)
 importFrom(ggplot2,aes_string)
-importFrom(ggplot2,element_text)
 importFrom(ggplot2,geom_abline)
 importFrom(ggplot2,geom_errorbar)
 importFrom(ggplot2,geom_point)
 importFrom(ggplot2,geom_smooth)
 importFrom(ggplot2,ggplot)
 importFrom(ggplot2,labs)
-importFrom(ggplot2,theme)
 importFrom(ggplot2,theme_bw)
 importFrom(hal9001,fit_hal)
 importFrom(haldensify,haldensify)
