### 2020-06-28 24020f2
fix plotting bugs: geom_abline extrapolation, weights; update docs
