+++ man/msm_vimshift.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/msm_vim.R
+% Please edit documentation in R/msm.R
 \name{msm_vimshift}
 \alias{msm_vimshift}
 \title{Working marginal structural model for causal effects of an intervention grid}
@@ -66,7 +66,7 @@ each parameter estimate separately. The default is to construct marginal
 confidence intervals for each parameter estimate rather than a simultaneous
 confidence band.}
 
-\item{...}{Additional arguments to be passed to \code{txshift}.}
+\item{...}{Additional arguments to be passed to \code{\link{txshift}}.}
 }
 \value{
 A \code{list} containing estimates of the individual counterfactual
