### 2020-07-07 #2222 @marcosvital
@nhejazi, could you please take a quick look on this version of the manuscript that I just generated? If everything seems ok with it, I'll start the reviewing process, and this will be the version that the reviewers will use to start it.
