### 2020-07-07 #2447 @whedon
[REVIEW]: txshift: Efficient estimation of the causal effects of stochastic interventions in R
**Submitting author:** @nhejazi (<a href="http://orcid.org/0000-0002-7127-2789">Nima Hejazi</a>)
**Repository:** <a href="https://github.com/nhejazi/txshift" target ="_blank">https://github.com/nhejazi/txshift</a>
**Version:** v0.3.4-joss
