### 2020-07-07 #2447 @marcosvital
@klmedeiros and @joethorley: thank you all for for accepting review this submission for JOSS.

Even if you are not starting the review right now, please accept the invite, as it has an expiration date (there is a link under Reviewer instructions & questions and you should also get an email notification). Furthermore, please check the instructions and checklists above, and let me know if you need any assistance.

You can also tag @nhejazi and @benkeser if you need to ask specific questions about the submission.
