### 2020-07-08 nhejazi/txshift#57 @joethorley
Please indicate how third parties can Seek support
https://github.com/openjournals/joss-reviews/issues/2447
