### 2020-07-08 nhejazi/txshift#57 @joethorley
The review asks in the Documentation section whether

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
