### 2020-07-08 nhejazi/txshift#57 @nhejazi
Thanks @joethorley for bringing up these issues. We had added https://github.com/nhejazi/txshift#issues and https://github.com/nhejazi/txshift#contributions (as well as https://github.com/nhejazi/txshift/blob/master/CONTRIBUTING.md) in order to address issues and contribution expectations for our users. Would you recommend documentation beyond these items? Happy to consider any such changes.
