### 2020-07-08 #2447 @joethorley
@marcosvital can you clarify what JOSS means by "clear guidelines for third parties wishing to Seek support".

See https://github.com/nhejazi/txshift/issues/57

I personally think the "Seek support" condition is covered by "Report issues or problems with the software" but please advise if this is not the case.
