### 2020-07-09 nhejazi/txshift#59 @joethorley
Can you provide examples for all the functions
> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
