### 2020-07-14 #2447 @marcosvital
Hi, @joethorley. I agree: having a clear way to report issues or problems is enough.

But now that you mentioned it, I think that a small modification could be helpful: users should be pointed to CONTRIBUTING.md, since there are detailed instructions on how to file an issue there. Currently, if someone reads the issues statement alone, it's not possible to know that there are more instructions under contributing guidelines. @nhejazi, does that make sense?
