### 2020-07-19 2cce363
add more guidelines on issues
