### 2020-07-19 #2447 @nhejazi
Hi @marcosvital and @joethorley, thanks for the discussion on report issues. I've made the change to point users to further instructions in the `CONTRIBUTING.md` in the "Issues" section of `README.Rmd` (see https://github.com/nhejazi/txshift/pull/60). I'll be using that PR to keep track of all revisions and will merge it once we've resolved all comments/issues raised in this thread.
