### 2020-08-18 nhejazi/txshift#59 @nhejazi
Thanks for raising this point. This package isn't intended to expose many user-facing functions - primarily, these are just `txshift()` and `msm_vimshift()`, alongside any methods exposed for the S3 classes these introduce (e.g., `summary`, `plot`). Following CRAN guidelines, we've provided examples for each of these functions and methods; however, there are a number of internal functions that are not intended to be user-facing.

We hope that this alleviates some of the concern around functions without examples in their documentation but would be happy to discuss any lingering concerns further.
