### 2020-08-18 nhejazi/txshift#59 @joethorley
@nhejazi - I saw the detailed documentation on various internal functions and assumed they were exported. You are quite right and I am closing this issue.
