### 2020-08-18 #2447 @joethorley
@nhejazi and @marcosvital - I've completed my review and I am of the opinion that the authors have satisfied all the requirements for publication in JOSS.
