### 2020-09-15 #2447 @klmedeiros
@nhejazi  My only reservations currently are that I could only get `txshift` to install by installing `haldensify ` from Github, not CRAN, which isn't documented anywhere.

Otherwise, this is excellent and I recommend for publication in JOSS.
