### 2020-09-15 #2447 @nhejazi
Thanks for the quick reply @klmedeiros. That's a great catch, I had missed that I had (at some point recently, I guess) bumped the version of the `haldensify` to 0.0.6 in `txshift` and forgotten to submit that minor release of `haldensify` to CRAN. I've just submitted that CRAN update and will ping here again once 0.0.6 of `haldensify` is on CRAN. This installation error should be taken care of then.
