### 2020-09-16 7675267
update pkgdown, refs, and tweak sl3 dep for CRAN
