+++ R/msm.R
@@ -58,42 +58,43 @@
 #'  these counterfactual means.
 #'
 #' @examples
-#' library(sl3)
-#' n_obs <- 100
-#' W <- as.numeric(replicate(1, rbinom(n_obs, 1, 0.5)))
-#' A <- as.numeric(rnorm(n_obs, mean = 2 * W, sd = 1))
-#' Y <- rbinom(n_obs, 1, plogis(2 * A - W))
-#' msm <- msm_vimshift(
-#'   W = W, A = A, Y = Y, estimator = "tmle",
-#'   g_fit_args = list(
-#'     fit_type = "sl",
-#'     sl_learners_density = Lrnr_density_hse$new(Lrnr_glm$new())
-#'   ),
-#'   Q_fit_args = list(
-#'     fit_type = "glm",
-#'     glm_formula = "Y ~ ."
-#'   ),
-#'   delta_grid = seq(-1, 1, 0.25)
-#' )
+#' if (require("sl3")) {
+#'   n_obs <- 100
+#'   W <- as.numeric(replicate(1, rbinom(n_obs, 1, 0.5)))
+#'   A <- as.numeric(rnorm(n_obs, mean = 2 * W, sd = 1))
+#'   Y <- rbinom(n_obs, 1, plogis(2 * A - W))
+#'   msm <- msm_vimshift(
+#'     W = W, A = A, Y = Y, estimator = "tmle",
+#'     g_fit_args = list(
+#'       fit_type = "sl",
+#'       sl_learners_density = Lrnr_density_hse$new(Lrnr_glm$new())
+#'     ),
+#'     Q_fit_args = list(
+#'       fit_type = "glm",
+#'       glm_formula = "Y ~ ."
+#'     ),
+#'     delta_grid = seq(-1, 1, 0.25)
+#'   )
 #'
-#' # fit a linear spline with knot at 0
-#' n_obs <- 100
-#' W <- as.numeric(replicate(1, rbinom(n_obs, 1, 0.5)))
-#' A <- as.numeric(rnorm(n_obs, mean = 2 * W, sd = 1))
-#' Y <- rbinom(n_obs, 1, plogis(0.1 * A * (A >= 0) - 3 * A * (A < 0) - W))
-#' msm <- msm_vimshift(
-#'   W = W, A = A, Y = Y, estimator = "tmle",
-#'   g_fit_args = list(
-#'     fit_type = "sl",
-#'     sl_learners_density = Lrnr_density_hse$new(Lrnr_glm$new())
-#'   ),
-#'   Q_fit_args = list(
-#'     fit_type = "glm",
-#'     glm_formula = "Y ~ ."
-#'   ),
-#'   delta_grid = seq(-1, 1, 0.25),
-#'   msm_form = list(type = "piecewise", knot = 0)
-#' )
+#'   # fit a linear spline with knot at 0
+#'   n_obs <- 100
+#'   W <- as.numeric(replicate(1, rbinom(n_obs, 1, 0.5)))
+#'   A <- as.numeric(rnorm(n_obs, mean = 2 * W, sd = 1))
+#'   Y <- rbinom(n_obs, 1, plogis(0.1 * A * (A >= 0) - 3 * A * (A < 0) - W))
+#'   msm <- msm_vimshift(
+#'     W = W, A = A, Y = Y, estimator = "tmle",
+#'     g_fit_args = list(
+#'       fit_type = "sl",
+#'       sl_learners_density = Lrnr_density_hse$new(Lrnr_glm$new())
+#'     ),
+#'     Q_fit_args = list(
+#'       fit_type = "glm",
+#'       glm_formula = "Y ~ ."
+#'     ),
+#'     delta_grid = seq(-1, 1, 0.25),
[…]
