+++ cran-comments.md
@@ -1,10 +1,9 @@
 ## Test environments
-* local Ubuntu 18.04: R 3.6.3 (stable)
+* local Ubuntu 20.04: R 4.0.2 (stable)
 * remote Ubuntu 16.04 on travis-ci:
-  * R 3.6.2 (stable)
-  * R 4.0.0 (under development)
-* local macOS install: R 3.6.1
-* Windows (on appveyor and winbuilder): R 3.6.3
+  * R 3.6.3 (old-stable)
+  * R 4.0.2 (stable)
+* Windows (on appveyor and winbuilder): R 4.0.2
 
 ## R CMD check results
 * There were 0 ERRORs.
@@ -16,3 +15,4 @@
 
 ## Additional Notes
 * This is a new CRAN submission.
+* It depends on `haldensify` v0.0.6, which was just submitted.
