### 2020-09-16 296b44f
fix links for CRAN
