### 2020-09-16 5ca1a2a
update cran comments
