### 2020-09-16 #2447 @nhejazi
Quick update: `haldensify` v0.0.6 is now available on CRAN, so the installation issue should be resolved. I've also made a few changes for CRAN submission of the package, which is now in process. If the review is complete, I'll go ahead and merge https://github.com/nhejazi/txshift/pull/60. Are we all set @joethorley @klmedeiros @marcosvital?
