### 2020-09-17 d659c8c
fixing cran initial checks
