+++ cran-comments.md
@@ -16,6 +16,7 @@
 ## Additional Notes
 * This is a new CRAN submission.
 * In the initial attempt, there were two significant issues:
-  * A few links throughout used `http`; these have been moved to `https`.
-  * The dependency `haldensify` v0.0.6 had just been submitted to CRAN but had
-    not been posted; it is now posted on CRAN.
+  * A few more links throughout used `http`; these have been moved to `https`
+    or otherwise changed entirely.
+  * The old URL `https://sl3.tlverse.org` has been replaced by
+    `https://tlverse.org/sl3`
