+++ inst/REFERENCES.bib
@@ -79,7 +79,7 @@
   author = {Hejazi, Nima S and {van der Laan}, Mark J and Janes, Holly E and
     Gilbert, Peter B and Benkeser, David C},
   year={2020},
-  url = {http://arxiv.org/abs/2003.13771},
+  url = {https://arxiv.org/abs/2003.13771},
   doi = {},
   journal = {Biometrics (Methodology)},
   volume={},
