### 2020-09-23 a65fcf6
minor updates for CRAN checks
