+++ vignettes/ipcw_txshift.Rmd
@@ -11,10 +11,6 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-```{r, echo=FALSE}
-options(scipen = 999)
-```
-
 ## Introduction
 
 For a more general introduction to the targeted maximum likelihood estimator
@@ -180,8 +176,9 @@ summary(tmle_hal_shift_2)
 
 To easily incorporate ensemble machine learning into the estimation procedure,
 we rely on the facilities provided in the [`sl3` R
-package](https://tlverse.org/sl3) [@coyle2020sl3]. For a complete guide on
-using the `sl3` R package, consider consulting https://tlverse.org/sl3.
+package](https://tlverse.org/sl3/) [@coyle2020sl3]. For a complete guide on
+using the `sl3` R package, consider consulting https://tlverse.org/sl3/ and
+https://tlverse.org/tlverse-handbook/sl3.html.
 
 ```{r make_sl, eval = FALSE}
 # SL learners to be used for most fits (e.g., IPCW, outcome regression)
@@ -204,7 +201,7 @@ sl_learner_density <- Lrnr_sl$new(learners = lrnr_dens_lib,
 
 ### Estimating the IPCW-TMLE with Optimal Stacked Regressions
 
-Using the framework provided by the [`sl3` package](https://tlverse.org/sl3),
+Using the framework provided by the [`sl3` package](https://tlverse.org/sl3/),
 the nuisance parameters of the TML estimator may be fit with ensemble learning,
 using the cross-validation framework of the Super Learner algorithm of
 @vdl2007super. In principal, it would be desirable to estimate the parameter
