### 2020-09-24 ab090d2
add a better start up message
