### 2020-09-25 84de261
update after CRAN release
