+++ paper/refs.bib
@@ -106,13 +106,26 @@
   author = {Hejazi, Nima S and {van der Laan}, Mark J and Janes, Holly E and
     Gilbert, Peter B and Benkeser, David C},
   year={2020},
-  url = {http://arxiv.org/abs/2003.13771},
-  doi = {},
-  journal = {Biometrics (Methodology)},
+  url = {https://doi.org/10.1111/biom.13375},
+  doi = {10.1111/biom.13375},
+  journal = {Biometrics},
   volume={},
   number={},
   pages={},
   publisher = {Wiley Online Library}
+}
+
+@article{hejazi2020hal9001,
+  title = {{hal9001}: Scalable highly adaptive lasso regression in {R}},
+  author = {Hejazi, Nima S and Coyle, Jeremy R and {van der Laan}, Mark J},
+  year  = {2020},
+  url = {},
+  doi = {},
+  journal = {Journal of Open Source Software},
+  volume = {},
+  number = {},
+  pages = {},
+  publisher = {The Open Journal}
 }
 
 @manual{R,
@@ -125,10 +138,10 @@
 }
 
 @manual{coyle2020sl3,
-  author = {Coyle, Jeremy R and Hejazi, Nima S and Malenica, Ivana and
-    Sofrygin, Oleg},
   title = {{sl3}: Modern Pipelines for Machine Learning and {Super
     Learning}},
+  author = {Coyle, Jeremy R and Hejazi, Nima S and Malenica, Ivana and
+    Sofrygin, Oleg},
   year = {2020},
   howpublished = {\url{https://github.com/tlverse/sl3}},
   note = {{R} package version 1.3.7},
@@ -137,8 +150,8 @@
 }
 
 @manual{coyle2020hal9001,
-  author = {Coyle, Jeremy R and Hejazi, Nima S and {van der Laan}, Mark J},
   title = {{hal9001}: The scalable highly adaptive lasso},
+  author = {Coyle, Jeremy R and Hejazi, Nima S and {van der Laan}, Mark J},
   year  = {2019},
   howpublished = {\url{https://CRAN.R-project.org/package=hal9001}},
   note = {{R} package version 0.2.6},
@@ -147,9 +160,9 @@
 }
 
 @manual{hejazi2020haldensify,
-  author = {Hejazi, Nima S and Benkeser, David C and {van der Laan}, Mark J},
   title = {{haldensify}: Conditional density estimation with the highly
     adaptive lasso},
+  author = {Hejazi, Nima S and Benkeser, David C and {van der Laan}, Mark J},
   year  = {2020},
   howpublished = {\url{https://CRAN.R-project.org/package=haldensify}},
   note = {{R} package version 0.0.5},
