### 2020-09-26 48415c1
update refs + add sl3 note
