+++ inst/REFERENCES.bib
@@ -79,13 +79,13 @@
   author = {Hejazi, Nima S and {van der Laan}, Mark J and Janes, Holly E and
     Gilbert, Peter B and Benkeser, David C},
   year={2020},
-  url = {https://arxiv.org/abs/2003.13771},
-  doi = {},
-  journal = {Biometrics (Methodology)},
+  journal = {Biometrics},
+  publisher = {Wiley Online Library},
   volume={},
   number={},
   pages={},
-  publisher = {Wiley Online Library}
+  url = {https://doi.org/10.1111/biom.13375},
+  doi = {10.1111/biom.13375}
 }
 
 @article{pfanzagl1985contributions,
@@ -114,7 +114,7 @@
   publisher={Cambridge university press}
 }
 
-@manual{coyle2020sl3,
+@software{coyle2020sl3,
   author = {Coyle, Jeremy R and Hejazi, Nima S and Malenica, Ivana and
     Sofrygin, Oleg},
   title = {{sl3}: Modern Pipelines for Machine Learning and {Super
@@ -126,12 +126,25 @@
   doi = {10.5281/zenodo.1342293}
 }
 
-@manual{coyle2020hal9001,
+@software{coyle2020hal9001-rpkg,
   author = {Coyle, Jeremy R and Hejazi, Nima S and {van der Laan}, Mark J},
   title = {{hal9001}: The scalable highly adaptive lasso},
   year  = {2020},
   howpublished = {\url{https://github.com/tlverse/hal9001}},
-  note = {{R} package version 0.2.6},
+  note = {{R} package version 0.2.7},
   url = {https://doi.org/10.5281/zenodo.3558313},
   doi = {10.5281/zenodo.3558313}
+}
+
+@article{hejazi2020hal9001-joss,
+  author = {Hejazi, Nima S and Coyle, Jeremy R and {van der Laan}, Mark J},
+  title = {{hal9001}: Scalable highly adaptive lasso regression in {R}},
+  year  = {2020},
+  journal = {Journal of Open Source Software},
+  publisher = {The Open Journal},
+  volume={5},
+  number={53},
+  pages={2526},
+  url = {https://doi.org/10.21105/joss.02526},
+  doi = {10.21105/joss.02526}
 }
