### 2020-09-28 f555e39
fix DOI + minor paper edits
