+++ paper/paper.md
@@ -6,7 +6,8 @@ tags:
   - two-phase sampling
   - efficient estimation
   - targeted learning
-  - stochastic interventions
+  - stochastic intervention
+  - modified treatment policy
   - R
 authors:
   - name: Nima S. Hejazi
@@ -22,7 +23,7 @@ affiliations:
     index: 2
   - name: Department of Biostatistics and Bioinformatics, Rollins School of Public Health, Emory University
     index: 3
-date: 25 September 2020
+date: 28 September 2020
 bibliography: refs.bib
 ---
 
@@ -30,14 +31,15 @@ bibliography: refs.bib
 
 The `txshift` `R` package aims to provide researchers in (bio)statistics,
 epidemiology, health policy, econometrics, and related disciplines with access
-to cutting-edge statistical methodology for evaluating the causal effects of
-_continuous-valued_ exposures. `txshift` estimates the causal effects of
-modified treatment policies (or "feasible interventions"), which take into
-account the natural value of an exposure in assigning an intervention level.
-What's more, the package provides independent corrections for estimating such
-effects under two-phase sampling (e.g., case-control) designs, allowing for the
-methodology to be readily applied in a diversity of real-world experimental and
-observational studies.
+to state-of-the-art statistical methodology for evaluating the causal effects of
+interventions on _continuous-valued_ exposures. `txshift` estimates the causal
+effects of modified treatment policies (or "feasible interventions"), which take
+into account the natural value of an exposure in assigning an intervention
+level. To accommodate use in experimental designs using two-phase sampling
+(e.g., case-control sampling), the package provides modern corrections for
+estimating causal effects subject to the limitations induced by these designs,
+allowing for the methodology to be readily applied in a diversity of real-world
+experimental and observational studies.
 
 # Background
 
@@ -76,22 +78,24 @@ a nonparametric working marginal structural model, analogous to a dose-response
 analysis. The `txshift` software package, for the `R` language and environment
 for statistical computing [@R], implements this methodology.
 
-`txshift` is designed to facilitate the simple construction of TML and one-step
+`txshift` is designed to facilitate the construction of TML and one-step
 estimators of the causal effects of modified treatment policies that shift the
-observed exposure value up (or down) by an arbitrary scalar $\delta$. The `R`
-package includes tools for deploying these efficient estimators under two-phase
-sampling designs, with two types of corrections: (1) a reweighting procedure
-that introduces inverse probability of censoring weights directly into an
-appropriate loss function, as discussed in @rose2011targeted2sd; as
-well as (2) a correction based on the efficient influence function, studied more
-thoroughly by @hejazi2020efficient. `txshift`
-integrates with the [`sl3` package](https://github.com/tlverse/sl3)
-[@coyle2020sl3] to allow for ensemble machine learning to be leveraged in the
-estimation of nuisance parameters. What's more, the `txshift` package draws on
-both the `hal9001` [@coyle2020hal9001; @hejazi2020hal9001] and `haldensify`
-[@hejazi2020haldensify] `R` packages  to allow each of the estimators to be
-constructed in a manner consistent with the methodological and theoretical
-advances of @hejazi2020efficient.
+observed exposure value up (or down) by an arbitrary scalar $\delta$, which may
+possibly take into account the natural value of the exposure. The `R` package
+includes tools for deploying these efficient estimators under two-phase sampling
+designs, with two types of corrections: (1) a reweighting procedure that
+introduces inverse probability of censoring weights directly into an appropriate
+loss function, as discussed in @rose2011targeted2sd; as well as (2) a correction
+based on the efficient influence function, studied more thoroughly by
[…]
