+++ paper/refs.bib
@@ -119,12 +119,12 @@
   title = {{hal9001}: Scalable highly adaptive lasso regression in {R}},
   author = {Hejazi, Nima S and Coyle, Jeremy R and {van der Laan}, Mark J},
   year  = {2020},
-  url = {},
-  doi = {},
+  url = {https://doi.org/10.21105/joss.02526},
+  doi = {10.21105/joss.02526},
   journal = {Journal of Open Source Software},
-  volume = {},
-  number = {},
-  pages = {},
+  volume = {5},
+  number = {53},
+  pages = {2526},
   publisher = {The Open Journal}
 }
 
