### 2020-09-28 #2447 @nhejazi
@marcosvital, no worries about the delay and thanks for catching the DOI issue. It's now fixed. Would you mind updating the package version in the first comment on this thread from 0.3.3 to 0.3.4? (I can't seem to do that myself.)
