### 2020-09-29 0d2b252
begin adding censoring node
