+++ R/txshift.R
@@ -10,16 +10,20 @@
 #'  baseline covariates.
 #' @param A A \code{numeric} vector corresponding to a treatment variable. The
 #'  parameter of interest is defined as a location shift of this quantity.
+#' @param C_cens A \code{numeric} indicator for whether a given observation was
+#'  subject to censoring by way of loss to follow-up. The default assumes no
+#'  censoring due to loss to follow-up.
 #' @param Y A \code{numeric} vector of the observed outcomes.
-#' @param C A \code{numeric} indicator for whether a given observation was
-#'  subject to censoring, used to compute an IPC-weighted estimator in cases
-#'  where two-stage sampling is performed. The default assumes no censoring.
+#' @param C_samp A \code{numeric} indicator for whether a given observation was
+#'  subject to censoring by being omitted from the second-stage sample, used to
+#'  compute an inverse probability of censoring weighted estimator in such
+#'  cases. The default assumes no censoring due to two-phase sampling.
 #' @param V The covariates that are used in determining the sampling procedure
 #'  that gives rise to censoring. The default is \code{NULL} and corresponds to
 #'  scenarios in which there is no censoring (in which case all values in the
-#'  preceding argument \code{C} must be uniquely 1). To specify this, pass in a
-#'  \code{character} vector identifying variables amongst W, A, Y thought to
-#'  have played a role in defining the sampling/censoring mechanism (C). This
+#'  preceding argument \code{C_samp} must be uniquely 1). To specify this, pass
+#'  in a \code{character} vector identifying variables amongst W, A, Y thought
+#'  to have impacted the definition of the sampling mechanism (C_samp). This
 #'  argument also accepts a \code{data.table} (or similar) object composed of
 #'  combinations of variables W, A, Y; use of this option is NOT recommended.
 #' @param delta A \code{numeric} value indicating the shift in the treatment to
@@ -40,7 +44,7 @@
 #'  or Super Learner ("sl"), and "external" a user-specified input of the form
 #'  produced by \code{\link{est_ipcw}}. NOTE THAT this first argument is not
 #'  passed to \code{\link{est_ipcw}}.
-#' @param g_fit_args A \code{list} of arguments, all but one of which are
+#' @param g_exp_fit_args A \code{list} of arguments, all but one of which are
 #'  passed to \code{\link{est_g}}. For details, consult the documentation of
 #'  \code{\link{est_g}}. The first element (i.e., \code{fit_type}) is used to
 #'  determine how this regression is fit: \code{"hal"} to estimate conditional
@@ -64,10 +68,10 @@
 #'  this to \code{"glm"} to instead use a simple linear regression model. In
 #'  this step, the efficient influence function (EIF) is regressed against
 #'  covariates contributing to the censoring mechanism (i.e., EIF ~ V | C = 1).
-#' @param ipcw_efficiency Whether to invoke an augmentation of the IPCW-TMLE
-#'  procedure that performs an iterative process to ensure efficiency of the
+#' @param ipcw_efficiency Whether to use an augmented inverse probability of
+#'  censoring weighted estimating equation to ensure efficiency of the
 #'  resulting estimate. The default is \code{TRUE}; only set to \code{FALSE} if
-#'  possible inefficiency of the IPCW-TMLE is not a concern.
[…]
