+++ README.Rmd
@@ -101,7 +101,7 @@ A <- rnorm(n_obs, mean = 2 * W, sd = 1)
 Y <- rbinom(n_obs, 1, plogis(A + W + rnorm(n_obs, mean = 0, sd = 1)))
 
 # now, let's introduce a a two-stage sampling process
-C <- rbinom(n_obs, 1, plogis(W + Y))
+C_samp <- rbinom(n_obs, 1, plogis(W + Y))
 
 # fit the full-data TMLE (ignoring two-phase sampling)
 tmle <- txshift(W = W, A = A, Y = Y, delta = 0.5,
@@ -129,7 +129,7 @@ summary(os)
 
 # fit an IPCW-TMLE to account for the two-phase sampling process
 ipcw_tmle <- txshift(W = W, A = A, Y = Y, delta = 0.5,
-                     C = C, V = c("W", "Y"),
+                     C_samp = C_samp, V = c("W", "Y"),
                      estimator = "tmle",
                      max_iter = 5,
                      ipcw_fit_args = list(fit_type = "glm"),
@@ -146,7 +146,7 @@ summary(ipcw_tmle)
 
 # compare with an IPCW-agumented one-step estimator under two-phase sampling
 ipcw_os <- txshift(W = W, A = A, Y = Y, delta = 0.5,
-                   C = C, V = c("W", "Y"),
+                   C_samp = C_samp, V = c("W", "Y"),
                    estimator = "onestep",
                    ipcw_fit_args = list(fit_type = "glm"),
                    g_fit_args = list(fit_type = "hal",
