### 2020-09-29 f3121c2
minor readability improvements for JOSS
