### 2020-09-29 #2447 @nhejazi
Thanks @marcosvital, I gave the paper another read through and made some minor corrections to improve its readability. I think I'm all set with it, but happy to correct any issues you might find. I'll rebuild the latest proof below.
