### 2020-09-29 9bfdd98
rm minor redundancy in JOSS paper
