### 2020-09-30 3477c25
introduce censoring estimation throughout
