+++ R/eifs.R
@@ -18,8 +18,8 @@ utils::globalVariables(c("..v_names"))
 #' @param estimator The type of estimator to be fit, either \code{"tmle"} for
 #'  targeted maximum likelihood estimation or \code{"onestep"} for a one-step
 #'  estimator.
-#' @param Delta Indicator for missingness. Used for compatibility with the
-#'  routine to compute IPCW-TML estimates.
+#' @param C_samp Indicator for missingness due to exclusion from second-stage
+#'  sample. Used for compatibility with the IPCW-TML estimation routine.
 #' @param ipc_weights A \code{numeric} vector that gives inverse probability of
 #'  censoring weights for each observation. These are generated by invoking the
 #'  routines for estimating the censoring mechanism.
@@ -41,7 +41,7 @@ eif <- function(Y,
                 Hn,
                 estimator = c("tmle", "onestep"),
                 fluc_mod_out = NULL,
-                Delta = rep(1, length(Y)),
+                C_samp = rep(1, length(Y)),
                 ipc_weights = rep(1, length(Y)),
                 ipc_weights_norm = rep(1, length(Y))) {
 
@@ -58,18 +58,18 @@ eif <- function(Y,
   }
 
   # compute substitution estimator of the stochastic shift parameter
-  param_obs_est <- rep(0, length(Delta))
-  param_obs_est[Delta == 1] <- ipc_weights_norm * Qn_shift
+  param_obs_est <- rep(0, length(C_samp))
+  param_obs_est[C_samp == 1] <- ipc_weights_norm * Qn_shift
   psi <- sum(param_obs_est)
 
   # compute the efficient influence function (EIF) / canonical gradient (D*)
-  eif <- rep(0, length(Delta))
-  eif[Delta == 1] <-
+  eif <- rep(0, length(C_samp))
+  eif[C_samp == 1] <-
     ipc_weights * (Hn$noshift * (Y - Qn_noshift) + (Qn_shift - psi))
 
   # this will be the outcome of the extra regression
-  eif_unweighted <- rep(0, length(Delta))
-  eif_unweighted[Delta == 1] <-
+  eif_unweighted <- rep(0, length(C_samp))
+  eif_unweighted[C_samp == 1] <-
     (Hn$noshift * (Y - Qn_noshift) + (Qn_shift - psi))
 
   # add mean of EIF to parameter estimate if fitting one-step
@@ -91,7 +91,7 @@ eif <- function(Y,
 
 ################################################################################
 
-#' Iterative IPCW Update Procedure of Efficient Influence Function
+#' Iterative IPCW Update Procedure of Augmented Efficient Influence Function
 #'
 #' @details An adaptation of the IPCW-TMLE for iteratively constructing an
 #'  efficient inverse probability of censoring weighted TML or one-step
@@ -102,14 +102,14 @@ eif <- function(Y,
 #' @param data_in A \code{data.table} containing variables and observations of
 #'  full data. That is, this corresponds to the data after application of a
 #'  censoring process.
-#' @param C A \code{numeric} binary vector giving the censoring status of a
-#'  given observation.
+#' @param C_samp A \code{numeric} indicator for missingness due to exclusion
+#'  the from second-stage sample.
 #' @param V A \code{data.table} giving the values across all observations of
 #'  all variables that play a role in the censoring mechanism.
 #' @param ipc_mech A \code{numeric} vector containing values that describe the
[…]
