### 2020-10-06 #2447 @marcosvital
Hi, @nhejazi! Almost there, but I would like to ask for two modifications:
 - please include the "statement of need" a separate paragraph, it should look [like this](https://github.com/openjournals/joss-papers/blob/joss.02584/joss.02584/10.21105.joss.02584.pdf) or [this](https://joss.theoj.org/papers/10.21105/joss.02342);
 - can you try start the summary with a more general, less technical explanation? JOSS guidelines for the summary asks for a "description of the high-level functionality and purpose of the software for a diverse, non-specialist audience". Usually, this can be easily achieved with one or two phrases (and I think that both papers linked above are good examples for that).
