### 2020-10-06 #2447 @nhejazi
Thanks, @marcosvital! I think the requested modifications are good points, and I've added both a very short introductory statement and a statement of need similar to those appearing in the examples. Please note that this package provides access to very sophisticated statistical methodology, so I'm not sure that I can write a more general introduction than the one now available.
