### 2020-10-06 #2447 @marcosvital
Looks great to me, @nhejazi, so let's proceed. You will need to archive the last release of the package (on Zenodo, figshare, or other) if you already didn't do this. After that, just let me know the archive DOI. Once this is done, we'll be ready to publish.
