### 2020-10-07 #2447 @nhejazi
Great! Thanks for the quick confirmation @marcosvital. I've created a new release at https://github.com/nhejazi/txshift/releases/tag/v0.3.4-joss and a Zenodo archive of that release at https://zenodo.org/record/4070043. The DOI for that release is 10.5281/zenodo.4070043.
