### 2020-10-07 #2447 @kthyng
Ok I have looked through and I see that the version is up to date, the zenodo archive exists and metadata matches the paper, and the paper reads through ok enough. We can proceed.
